refactor: changed filenames, deleted unnecessary code
