+++ src/main.rs
@@ -4,7 +4,7 @@ mod conversions;
 mod defines;
 mod logging;
 mod lora;
-mod lora_trait;
+// mod lora_trait;
 mod mqtt;
 mod packet;
 mod post;
@@ -22,8 +22,7 @@ use crate::sx1278::SX1278;
 use bme280::BME280Sensor;
 use log::{error, info};
 // use lora::LoRa;
-use crate::lora_trait::LoRa;
-use lora::start_lora;
+use lora::{start_lora, LoRa};
 use mqtt::{MQTTMessage, Mqtt};
 use packet::Status;
 use std::env;
